Minor tidy ups on PoolConfigurationModule tests
